Reorganize the Slack bug-scan worker into its own package
Moves the parts of the Slack bug-scan feature that do not depend on
execution-engine's worker runtime or app's plan-submission internals
into a new @invoker/slack-bug-scan package: the client/classifier/
plan-submitter contract, the fingerprint/prefilter/repo-url logic,
the tuning env resolvers, and the concrete Slack and Anthropic
implementations.

The worker's execution-engine wrapper (createWorkerRuntime, the
ledger-backed scan loop) and the app's plan drafter (worktree
provisioning, plan submission) stay where they are, since both
genuinely depend on their layer's own internals; they now import the
moved pieces from the new package instead of local files. The new
package's contract.ts additionally consolidates four small type/
interface definitions (classifier and plan-submitter shapes, and the
worker-kind constant) that previously lived inline in the
execution-engine wrapper, since they describe the same plugin surface
as the client interface and had no execution-engine dependency of
their own. No other logic changed.

@slack/web-api moves from a runtime dependency of packages/app to a
dev-only dependency: nothing in packages/app's runtime code imports it
directly anymore now that the real Slack client moved out, but a
later commit's test still needs it to mock the exact module instance
that client resolves.
