+++ packages/app/package.json
@@ -78,16 +78,14 @@
     "@invoker/runtime-adapters": "workspace:*",
     "@invoker/runtime-service": "workspace:*",
     "@invoker/shell": "workspace:*",
+    "@invoker/slack-bug-scan": "workspace:*",
     "@invoker/surfaces": "workspace:*",
     "@invoker/transport": "workspace:*",
     "@invoker/workflow-core": "workspace:*",
-    "@anthropic-ai/sdk": "^0.115.0",
-    "@slack/web-api": "^7.14.1",
     "dockerode": "^4.0.0",
     "sql.js": "^1.12.0",
     "dotenv": "^16.0.0",
-    "yaml": "^2.7.0",
-    "zod": "^4.4.3"
+    "yaml": "^2.7.0"
   },
   "optionalDependencies": {
     "node-pty": "^1.1.0"
@@ -96,6 +94,7 @@
     "@electron/rebuild": "^4.0.3",
     "@invoker/test-kit": "workspace:*",
     "@playwright/test": "^1.58.2",
+    "@slack/web-api": "^7.14.1",
     "electron": "^35.0.0",
     "tsup": "^8.0.0",
     "typescript": "^5.7.0",
