+++ packages/execution-engine/src/index.ts
@@ -77,12 +77,7 @@ export * from './workers/reaper-worker.js';
 export * from './workers/requeue-worker.js';
 export * from './workers/workflow-resume-worker.js';
 export * from './workers/slack-bug-scan-worker.js';
-export * from './workers/slack-bug-scan-client.js';
-export * from './workers/slack-bug-scan-repo-url.js';
-export * from './workers/slack-bug-scan-fingerprint.js';
-export * from './workers/slack-bug-scan-prefilter.js';
 export * from './workers/slack-bug-scan-scanner.js';
-export * from './workers/slack-bug-scan.js';
 export * from './reconcile-terminal-worker-actions.js';
 export * from './requeue-attempt-ledger.js';
 export * from './auto-fix-gating.js';
