+++ packages/slack-bug-scan/src/contract.ts
@@ -1,3 +1,5 @@
+export const SLACK_BUG_SCAN_WORKER_KIND = 'slack-bug-scan';
+
 export interface SlackBugScanChannelSummary {
   id: string;
   name?: string;
@@ -20,3 +22,28 @@ export interface SlackBugScanClient {
   listReplies(channelId: string, threadTs: string): Promise<SlackBugScanMessage[]>;
   postMessage(channelId: string, threadTs: string, text: string): Promise<{ ts: string }>;
 }
+
+export interface SlackBugScanClassifyResult {
+  isBugComplaint: boolean;
+  problemStatement?: string;
+}
+
+export type SlackBugScanClassifier = (input: {
+  channelId: string;
+  threadTs: string;
+  repoUrl: string;
+  threadText: string;
+}) => Promise<SlackBugScanClassifyResult>;
+
+export interface SlackBugScanDraftResult {
+  planName: string;
+  workflowId: string;
+}
+
+export type SlackBugScanPlanSubmitter = (input: {
+  channelId: string;
+  threadTs: string;
+  repoUrl: string;
+  problemStatement: string;
+  threadText: string;
+}) => Promise<SlackBugScanDraftResult>;
