+++ packages/slack-bug-scan/src/fingerprint.ts
