+++ packages/slack-bug-scan/src/limits.ts
