+++ packages/slack-bug-scan/src/prefilter.ts
